Test invalid Lods in the textureSize tests
The spec is being updated to say that these give undefined values, not
undefined results, so add some tests that such lods don't crash drivers.
The results for these tests aren't verified, but they could be if a
feature is added to define these values in the future. (That's currently
set for maintenance4)

Components: Vulkan
Affects: dEQP-VK.glsl.texture_functions.query.texturesize.*

Change-Id: I5143c5865df6a78a5feab8ee985ca739602b4c6b