Avoid querying YCbCr combined sampler count in RGBA
Several YCbCr query tests try to obtain the value of
combinedImageSamplerDescriptorCount for RGBA formats without checking if
the implementation actually supports VK_KHR_sampler_ycbcr_conversion or
the samplerYcbcrConversion feature.

On the one hand, obtaining that value for RGBA formats may not be
completely right due to the spec description supposing the format would
be a multiplanar format.

On the other hand, if the implementation conforms to Vulkan 1.0 and does
not support the extension, the structure will be ignored and
combinedImageSamplerDescriptorCount will be zero, resulting in wrong
behavior when creating a descriptor pool and allocating a descriptor
set from it.

Affects:
dEQP-VK.ycbcr.query.*r8g8b8a8_unorm*

Components: Vulkan
VK-GL-CTS Issue: 2724

Change-Id: I8660ccd5bdbb2e12133f76855049e9efcacc7888