Move and rename descriptorset_random tests
Change 916d8734f509267b4999a1009e78c2beb43eed15 involved a large
overhaul of the descriptorset_random tests. This led to a much larger
overall set of such tests, but such that there is no direct
correspondence between the old test names and the new test
names. Furthermore, all the old tests were removed from the
vk-master-2019.txt and vk-master-2020.txt test lists, and all of the
new tests were added to the vk-master-2019.txt test list.

This change is a best effort to:

- Revert the changes to the 2019 and 2020 test lists so that each
  contains the same number of descriptorset_random tests that it
  contained prior to 916d8734f509267b4999a1009e78c2beb43eed15;

- Ensure that the descriptorset_random tests that appear in the 2019
  list map as closely as possible to corresponding updated
  descriptorset_random tests, and similarly for the 2020 test list.

Deciding how to map an old, removed test T to some new test T' was
done as follows:

1. Compute the longest prefix of T such that at least one replacement
   test starts with that prefix. Call the set of tests with this
   prefix Candidates.

2. Choose T' to be the member of candidates such that T and T' have
   the longest common suffix.

3. Remove T' from the set of available new tests, so that a future
   test name will not also be mapped to T'.

This should lead to the revised 2019 and 2020 lists containing
descriptorset_random tests that achieve around the same coverage that
was achieved by the tests prior to change
916d8734f509267b4999a1009e78c2beb43eed15.

Bug: 169021525

Change-Id: I890ed52265ea04b7ec57e6aa4d8a1b3c450460dc