Fix decoration_group tests
Specification states that: Variable decorated with Invariant must
only be identified with the Input or Output storage class in Vulkan
environment. Purpose of the test wasn't to test invariance, but to
use three decorations on group3. This change replaces Invariant
with Flat decoration so that validation error is not triggered.

Components: Vulkan

Affects:
dEQP-VK.spirv_assembly.instruction.graphics.decoration_group.*

Change-Id: I460b54988f2991e007fee27f1e96c39dbbcec74e