Terminate with an error if unrecognized arguments are passed to dEQP executable
de::cmdline seems to support options (starting with '-' or '--') and generic
arguments.

It seems the command line interface of dEQP, which wraps de:cmdline,  was
not making any use of the arguments, but at the same time ignored them.
This made a possibility for unrecognized command line arguments to be
completely ignored, instead of throwing and error

Components: Framework

VK-GL-CTS Issue: 2727

Change-Id: Ibea7413edec2899d33484ac8bb8183c4f4d398ac