Change format used in pipeline cache control tests
The format used as a colour attachment by the pipeline creation cache
control tests is not required to be supported as a color attachment and
the tests do not query for it being supported.

Change it to R8G8B8A8 which is required to be supported as the specific
format is not important for this test.

Affects:
dEQP-VK.pipeline.creation_cache_control.graphics_pipelines.*

Components: Vulkan

VK-GL-CTS issue: 2856

Change-Id: I7ecfc4de8584dba6824101cb717313247070f12f