Update Amber
Amber was causing validation errors in some tests because VK_WHOLE_SIZE
as the buffer range and dynamic offsets were used together in
uniform/storage buffer bindings, which is not allowed.

This issue is fixed in the latest version of Amber.

VK-GL-CTS issue: 2873

Affects: All Amber tests

Components: Vulkan, Framework
Change-Id: I79a003adeebc4718b1c794573fa476a4288f17c5