Don't run _concurrent sync tests on 1 queue devices
The tests are testing synchronisation primitives in combination with
images marked as SHARING_MODE_CONCURRENT but this is not possible on
deivces that only support a single queue family. Because creating an
image with concurrent sharing requires two distinct queue families, it
is not supported on these devices.

Component: Vulkan
Affects: dEQP-VK.synchronization*.multi_queue.*_concurrent

Change-Id: I30a05bb085eef9e3087adfa12377a4909942f9c3