Skip unsupported allocations in invariance test
dEQP-VK.api.invariance.random is a test that performs a high number of
memory allocations (1000 as of this commit). Some of them may throw a
not supported exception that wasn't being caught and would result in the
whole test being marked as unsupported.

This change catches the exception and only returns "not supported" if
none of the allocations were supported. Otherwise, it runs all possible
allocations and checks results for the supported ones.

Affected tests:
dEQP-VK.api.invariance.random

Components: Vulkan
VK-GL-CTS issue: 2110

Change-Id: I92163c42e33671291dd9fc0c753dc43b4a058796