Retry lesser FBO sample counts if UNSUPPORTED
On some of our implementations we return GL_FRAMEBUFFER_UNSUPPORTED
when the no-attachment maximum width, height and samples are all
used in combination. This is because the combination goes beyond
some of our hardware limits. It would be fine to use either
maximum width/height or maximum samples, but not all
together.

Since GL_FRAMEBUFFER_UNSUPPORTED is a valid FBO status for
implementation-defined reasons, the test should recognize this
and try again with lesser limits until the FBO is complete.

Affects:

dEQP-GL45.functional.fbo.no_attachments.maximums.all

Components: OpenGL

VK-GL-CTS issue: 2973

Change-Id: I8db2e9f802b3f4d1ee59bbfbfa1935bbe3cc6d00