Remove extended image format capabilities from more tests
The extended storage formats capability is mandatory, so simplify the
code by using ti all the time, instead of trying to determine when it is
required.

Also, change the method of determining when 64 bit support is required
to be consistent with the image.extend_operands_spirv1p4 tests.

Components: Framework, Vulkan
Affects: dEQP-VK.image.mismatched_write_op.*

Change-Id: Ia9cc474977e71a8682c886f2f6cd852be38f06c4