Use common code for robustness2 tests
Where the code is the same for every stage, keep a single copy instead
of writing it out every time. This makes it easier to understand what is
different between the cases and why. It also means that we only have to
fix issues once instead of 6 times.

Components: Vulkan
Affects: dEQP-VK.robustness.robustness2.*
         dEQP-VK.robustness.image_robustness.*

Change-Id: I8dcf6efee86165f78f657358a0f11fd6e2e2d163