Update checking of VulkanFeatures structs
The spirv_assembly tests defined various helpers for checking whether
certain features are supported but they worked inconsistently and were
repeated in various places. This change pulls the checking into a single
function which can be used to check instead of having to call a while
set of functions to achieve the same end.

The higher-level checking functions have always masked out a subset of
the features that are requested, and this is not changed, so requests
for certain things that look correct will silently not check all the
features.

The extension feature enums that were defined just for the checking
system are replaced with the standard structs. They're easier to use
because they don't need to be defined afresh when extensions are added
here, and because the names are automatically consistent with the names
that are known from the API. In at least 1 case the existing names were
pretty hard to decode correctly.

Components: Vulkan
Affects: dEQP-VK.spirv_assembly.*

Change-Id: Icd2fc3f7d37e0ec3cfed9a80f3152617c8f599a3