Issue in serial storage address computation
When serial storage and src/dst buildTypes are different, serial storage address computation is not correct. In the beginning of the test case, serial storage buildType is either set to host or device according to test case being run (cpu-built or gpu-built). 
In the 2nd phase of this test case (deserialize all from the previous step to a new top-level AS) VkCopyMemoryToAccelerationStructureInfoKHR is filled with srcAddress according to testCase buildType, and dst acceleration structure is device.  
In this stage, if testCase builtType is device, the serialStorage address computation is correct. However, if testCase buildType is host then the issue arises. The address is computed according to host, although device address is needed.

That is, for serial storage address computation: buildType of the acceleration structure (src or dst) should be used rather than serial storage buildType which is initially set according to test case being run.

Affects:
dEQP-VK.ray_tracing_pipeline.acceleration_structures.header_bottom_address.*
 
Components: Vulkan
VK-GL-CTS issue: 3028

Change-Id: If8c0387862ce6f1db00fc61016dd0e602ff72ba8