Test non-Sample Shading FragCoord.xy
In these test we are testing that FragCoord reflect the location of the
center of the fragment when Sample Shading is disabled. There is also a
set of tests which are using Centroid interpolation decoration on
FragCoord.

Components: Vulkan

VK-GL-CTS Issue: 3040

New tests: dEQP-VK.glsl.builtin_var.fragcoord_msaa_input.*no_sample*

Change-Id: Ic84bd57d8118e001bcd9e52ecb2e4193ff332ba6