Fix layout transitions in dynamic_rendering tests
With dynamic rendering, we no longer have initial and final layout
transitions provided automatically as part of vkCmdBeginRendering() and
vkCmdEndRendering().  Instead, it's the responsibility of the client to
transition the images into and out of the layout passed to
vkCmdBeginRendering().  When translating from Vulkan 1.0 style
subpasses, this means we need to do a pipeline barrier before and after
rendering.

Affects:
dEQP-VK.renderpass_with_dynamic_rendering.*

Change-Id: I56214198a8b3d81e3030164de045b8ca03a44f88