Fix invocation count verification
The shaders in these tests track the number of invocations executed and
verify that against the expected number of invocations for the test.

The spec only guarentees that each fragment is executed at least once.
So an implementation is free to execute each fragment multiple times
if required (for example, an implementation might required a shader to
be run at sample rate).

Change this verification to only check against the minmum number of
invocations.

Affects: dEQP-VK.fragment_operations.early_fragment.samplemask_*

Components: Vulkan

VK-GL-CTS issue: 3278

Change-Id: I93f1aa42dd6a13486bed044e361671e5a2f2cb07