Test no-op image layout transitions in VK_KHR_synchronization2
VK_KHR_synchronization2 allows applications to specify
oldLayout==newLayout in image barriers to skip the layout transition.
It's also allowed when oldLayout and newLayout are both UNDEFINED even
though the normal behaviour for UNDEFINED -> ANYTHING is to perform
a discard.

This commit adds a test which ensures that image contents are preserved
when doing a transition like this.

VK-GL-CTS issue: 3045

New Tests:
dEQP-VK.synchronization2.layout_transition.no_op

Components: Vulkan
Change-Id: I022ca662ad50e910d7813713e591f2f6f9600d1f