Query default FBO Depth, Stencil only when format requires
Earlier commit removed DSA usage but introduced a problem with
framebuffer parameter type and size queries in some configs.

Without DSA, we get errors if querying for attachment that does
not exist in framebuffer. Patch introduces a check to make sure
we do not query for separate depth/stencil attachment if that
would not exist in format.

VK-GL-CTS issue: 3092

Affects:
KHR-*.internalformat.renderbuffer.*

Signed-off-by: Tapani Pälli <tapani.palli@intel.com>
Change-Id: I65ba7f74e5abd3a5bff17cd74ca4abcdf3f0ab53