Remove time measurement on deferred host build tests
The time measurement is wrongly implemented in this test
- it measures also thread creation, allocations and VK setup time, additionally
to AS build.  It is not easily fixable.

Moreover, the spec says that offloading the work to background threads
is 'desirable', however it never says it is faster.

VK-GL-CTS Issue: 3435

Components: vulkan

Affects:
dEQP-VK.ray_tracing_pipeline.acceleration_structures.*
dEQP-VK.ray_query.acceleration_structures.*

Change-Id: I41d836e2e454b9fb9bb46109edddd4eda19370e9