Fix load_store_op_none dynamic rendering validation error
The test case below was using a multisample color attachment with
integer format and specified VK_RESOLVE_MODE_AVERAGE_BIT as the resolve
mode, which is illegal. The purpose of the test is different, and the
resolve mode is not relevant, so it can be switched to one of the legal
values.

Affected tests:
dEQP-VK.dynamic_rendering.suballocation.load_store_op_none.color_load_op_none_store_op_none_resolve

Components: Vulkan
VK-GL-CTS issue: 3509

Change-Id: I42cf959f67d80694aee37c19c966ab037609d44f