Actually test compatible depth/stencil attachment formats
These tests claimed to test "compatible" depth resolve formats, i.e.,
resolving to a depth-only format from a depth/stencil format.  However,
it only set the depth-only format on the renderpass and not on the
single-sampled image.  This meant it was in violation of the following
VU because the image view and attachment format didn't match:

    VUID-VkRenderPassBeginInfo-framebuffer-03216

    If framebuffer was created with a VkFramebufferCreateInfo::flags
    value that included VK_FRAMEBUFFER_CREATE_IMAGELESS_BIT, each
    element of the pAttachments member of a
    VkRenderPassAttachmentBeginInfo structure included in the pNext
    chain must be a VkImageView of an image created with a value of
    VkImageViewCreateInfo::format equal to the corresponding value of
    VkAttachmentDescription::format in renderPass

Unfortunately, once I fixed the format, it turned into a bit of a rabbit
hole because aspects were also messed up various places and then layouts
as well.

Components: Vulkan
Affects: dEQP-VK.renderpass2.depth_stencil_resolve.*separate_layouts.*

Change-Id: I34f5dd2cbccae2a695cbe05a22908ba544b0a6a6