Fix multiview dynamic rendering tests to enable dynamicRendering feature
The test uses own, custom VkDevice which is created without enabling
'dynamicRendering' feature -  while this feature is used by the test.

The fix is to add VkPhysicalDeviceDynamicRenderingFeatures to the
chain of create infos used for getting and requesting features, so
the feature would be alwyas enabled if supported.

(The support of the feature is checked elsewhere).

Components: Vulkan

VK-GL-CTS Issue 3537

Affects: dEQP-VK.multiview.dynamic_rendering,*

Change-Id: Ibcc9a27e01c5deeff1ef90baf97eee5148244603