Fix regression from CL-9387 to handle empty cache properly
The new program cache code from CL-9387 didn't handle the case where
cacheSearch() is called with an empty cache. It would look at the
first node key regardless, which is uninitialized data, and
attempt to follow the left/right node pointers, which were
also uninitialized. This could result in crash or worse.

This CL simply checks if the cache is empty, and if so doesn't
attempt to read the invalid first node.

Affects:
* (all tests with shaders)

VK-GL-CTS issue: 3565

Components: Vulkan, Framework
Change-Id: Ia813d8caf3170b0c1995bdc43a5f345b578ab615