Fix incorrect staging buffer size calculation
The dEQP-VK.api.image_clearing.* tests use a staging buffer to
pre-clear the images to a known value before each test. The size
of this buffer was previously determined by taking the size from the
VkMemoryRequirements obtained for the image. This fails if the size
required to store the image is less than the size required to store
the raw texel data in the buffer, which is later to be copied into
the image. This commit changes that and instead computes the size
of the staging buffer based on image extent, format and number of
layers.

Affects:
dEQP-VK.api.image_clearing.*

Components: Vulkan

VK-GL-CTS issue: 3809

Change-Id: I2b74697d54e11ddd7598299d57d02205bd2571d5