Support commandPoolResetCommandBuffer==VK_FALSE for VulkanSC
When this property is false, resetting command buffers (either
explicitly or implicitly via vkBeginCommandBuffer) is not supported,
and the only way to reset is to reset the whole pool.

Add a helper method to the Context to reset a pool if needed, and call
it everywhere it's needed (at least, in tests that NVIDIA's
implementation supports).

Also add a missing VK_CHECK around calls to
vk{Reset,Begin}CommandBuffer.

Affects: dEQP-VKSC.*

Components: VulkanSC

VK-GL-CTS Issue: 3837

Change-Id: Ia2f80997103881c9b644b5a8124bef5d41ab5a43