Enable extended_dynamic_state tests for VK1.3
The partial promotion of the 2 extended dynamic state extensions meant
that the tests were not being run on VK 1.3 core if the extension was
not also exposed.

This change has the tests checkSupport always based off the feature
struct, with the framework extended to fill in the extension features on
core implementations. This is consistent with how unsupported extensions
are normally handled (the framework fills in an empty feature struct, so
they can always be used to determine test support).

Components: Vulkan, Framework
VK-GL-CTS issue: 3937
Affects: dEQP-VK.pipeline.*.extended_dynamic_state.*

Change-Id: Ifc3e821a50e55e54cc0bbd938d2080e526a7b55f