Limit call of getQueryPoolResults to BUILD_TYPE_DEVICE
When the buildType is BUILD_TYPE_HOST, queryPool results are retrieved immediately to the CPU, and queryPool is not set to the result value.
Due to unset value at queryPool, later calling getQueryPoolResults causes the size to be messed up, and assertion to be trigerred.
The intention to call getQueryPoolResults is for BUILD_TYPE_DEVICE where you need to wait for cmdBuffer to be executed.
Call of getQueryPoolResults should only be made if the build type is BUILD_TYPE_DEVICE.

Affects:
dEQP-VK.ray_tracing_pipeline.acceleration_structures.copy_within_pipeline.cpu.*

Components: Vulkan

https://gitlab.khronos.org/Tracker/vk-gl-cts/-/issues/4016

Change-Id: I2e8c3dc9c56c98c288bba17d6c4074a9b88e0851