Make application_parameters tests fail without vendor-specific info
VK-GL-CTS issue: 4018

The "invalid_param_value" and "valid" test variants need vendor-specific
values to be added to the test case. Previously this reported "unsupported"
if the necessary values aren't added. It was believe this could be easily
missed by vendors, so changing this to an error as agreed.

Note that with this change, if an implementation supportsi
VK_EXT_application_parameters and does nothing else - there will be 4 failures.

Component: VulkanSC
Affects: dEQP-VKSC.sc.application_parameters.*

Change-Id: I76a1901cc15e0966ffdd00a817aebbeb72f7d019