Use correct extent for 1D compute texel_view_compatible tests
We weren't sanitizing the size in the parameters, resulting in
copyDataToImage() passing an extent with a non-1 height to when
uploading to a 1D image. Sanitize fullSize, and remove a case where
where we were workaround it not being sanitized.

Components: Vulkan

Affects: dEQP-VK.image.texel_view_compatible.compute.basic.1d_image.*

VK-GL-CTS Issue: 4143

Change-Id: Icdb50c09137a947f47b1afc7426e53e73c31f689