Add VK_NULL_HANDLE definition to vkDefs.hpp
VK_NULL_HANDLE was notably missing for usage in CTS tests, and in most
of them it was being replaced by DE_NULL, which in the past also doubled
as NULL and nullptr.

Since VK-GL-CTS was updated to support C++11, nullptr was available as a
more natural way of using null pointers, but DE_NULL still had to be
used as the VK_NULL_HANDLE replacement.

This commit defines VK_NULL_HANDLE to be DE_NULL, which is known to
work, and makes its definition available globally as expected.

No test results should be affected.

VK-GL-CTS issue: 1701
Components: Vulkan

Change-Id: I188ab0ae8764ba25be38048479d62831909fbe29