Detect fallback texture failures in GL linear border_color tests
The 256x256 texture being rendered to the middle 1/3 of the screen
should mostly come through as-is, with just the pixel or so of samples
on each edge blending with the border color.

This catches rendering failure with ANGLE+radv that was not detected,
but I noticed when comparing with the rendering failure that the test
caught on ANGLE+anv.

ANGLE is providing a fallback texture on 2 of the tests in this group
on these chips (Texture2DDC16Linear and Texture2DDC32FLinear), for
reasons I have yet to debug, so the rendering magnifies a 1x1 texture
(so only approximately the center point is black, and the rest is
blended with red in every direction).  This only happened to fail on
anv because the loop walking to the top is off by one and starts with
comparing a pixel *below* the center point to the center, and on that
driver the pixel is slightly blended with red.

Change-Id: Ia9ea6d6e08f55547a843cf223a643a89d04f4825
Components: OpenGL, OpenGL ES
Affected tests:
KHR-GLES31.core.texture_border_clamp.*
KHR-GL*.texture_border_clamp.*