Fix bind-transfer synchronization in sparse texture tests
The copyBufferToImage() utility function submits a command buffer with a
copy-buffer-to-image transfer operation. The caller can optionally pass
a semaphore for the transfer to wait on, but the semaphore's wait flags
are set to VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT, which does not include
transfer operations.

As a consequence of this, the transfer operation may ignore the wait
semaphore. In the affected test set this means that the transfer may
happen before sparse bind operations are completed.

Affects: dEQP-VK.texture.*sparse*

Components: Vulkan

VK-GL-CTS issue: 4183

Change-Id: Ia0f0c721e492703301440dbfcf07ed6495fddb1e