Do not allow to call vkSetEvent after submission, which is invalid.
The spec says:

"If a command buffer submission waits for any events to be signaled,
the application must ensure that command buffer submissions will be
able to complete without any subsequent operations by the application.
Events signaled by the host must be signaled before the command buffer
waits on those events."

This means it's not valid to signal an event on the host after submission
of a command buffer containing a wait for the event. Also it forces the
command buffer to stop for an unknown amount of time, as much as the time
needed for the host to signal the event with vkSetEvent.

However there were some CTS tests doing like this, which should be
fixed.

For vktApiExternalMemoryTests.cpp,
There was a vkSetEvent call after submission to hold a command to be
executed before signaling to get a native file descriptor, which is
not valid. So the tests should allow to get -1. Also we do an atomic calculation
to avoid singalling right after submit when trying to get a native fd.
The simple atomic calculation is evaluated to take about 10ms.

For vktApiExternalMemoryTests.cpp and vktMemoryExternalMemoryHostTests.cpp,
replace vkSetEvent with timeline semaphore to signal from the host.

Affects:
dEQP-VK.api.external.semaphore.*
dEQP-VK.api.external.fence.*
dEQP-VK.api.command_buffers.secondary_execute_twice
dEQP-VK.memory.external_memory_host.synchronization.synchronization

VK-GL-CTS Issue: 4075

Components: Vulkan
Change-Id: I5544bb15a418bbbbd6552ca04bec673a56ca3109