Convert ray gen. robustness tests from NV to KHR
This commit mainly changes the robustness tests that run on the ray
generation shading stage from the NVIDIA extension to the Khronos one.

In addition, it fixes and simplifies the number and ways robustness
devices are created for these tests, reducing the number and variants of
them, and making sure the right device interface is used with the right
device.

Affects:
dEQP-VK.robustness.*

Components: Vulkan
VK-GL-CTS issue: 4235

Change-Id: Ib91ac9cfe22b1ceb9d4323e2f92bae87d5659b3b