Pass fragment shading rate info to pre rasterization shader state
[Note]
This is a backporting patch, that landed already on the branch 1.3.4
and main. But the issue still exist on the branch 1.3.2/1.3.3, so I
create this CL that needs to propagate only to 1.3.3.

Fragment shading rate info needs to be passed both as part of the
pre-rasterization shader state and as part of the fragment shader state
when creating pipelines, and it needs to be consistent in both cases.

The graphics pipeline library utilities were only passing that
information as part of the fragment shader state.

Affects:
dEQP-VK.fragment_shading_rate.*.misc_tests.*
dEQP-VK.fragment_shading_rate.*.basic.*
dEQP-VK.pipeline.*.mixed_attachment_samples.*
dEQP-VK.*.sample_locations_ext.*
dEQP-VK.pipeline.*.multisample_with_fragment_shading_rate.*

Components: Vulkan

VK-GL-CTS issue: 4006

Change-Id: I4c01a7836ed45f513f166e787d1e709c9a612018