Try to inform the compiler when dynamically changing the rounding mode.
Changing the dynamic rounding mode without informing the compiler
we're intending to do that is undefined, but unfortunately the
compilers make this really annoying to fix in a portable way.

Clang 16 has started performing optimizations in this area of
undefined behaviour, causing failures in release builds for,

dEQP-VK.glsl.builtin.precision.fract.highp.*

C standard states,

> If part of a program tests floating-point status flags, sets
> floating-point control modes, or runs under non-default mode
> settings, but was translated with the state for the FENV_ACCESS
> pragma ‘‘off’’, the behavior is undefined.

Adding to the confusion is the fact that

> The default state (‘‘on’’ or ‘‘off’’) for the pragma is
implementation-defined.

Of GCC, Clang and MSC, the default state is OFF, but GCC doesn't
actually honour that pragma, so you can't change the default with
GCC. This has been a problem with GCC for 15 years
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=34678

This patch at least fixes the observed issue with Clang 16, and to
inform MSC we are dynamically changing the rounding mode so it doesn't
do anything funky in the future. There doesn't seem to be anything we
can do with GCC, apart from rearranging the code on a case-by-case
basis, using volatile modifiers or compiler barriers to force it not
to perform any code motion in these areas.

Places where the rounding mode was dynamically changed have been
marked up to warn the compiler when possible against doing code motion
which technically violates the strict FP profile.

The GPU shader FP64 test was redundantly setting the rounding mode,
DE_ROUNDINGMODE_TO_NEAREST_EVEN is the default mode, and also ensured
to be the rounding mode by tcuApp at startup.

Affects:

KHR-GL*gpu_shader5.fma_precision*
KHR-GL*.gpu_shader_fp64*
dEQP-VK.glsl.builtin.precision*

Components: Vulkan, OpenGL, Framework

VK-GL-CTS issue: 4285

Change-Id: I70fd11ce36f85571dd5cd3f7541d32777609cc4c