Show dEQP version even if building from git
When building from git, the git commit hash was displayed twice.
Instead, let's use the output of `git describe`, which looks like this:
  vulkan-cts-1.3.5.0-163-g20d674342f008624b82e

The various outputs and their format are explained in the git docs [1],
but the important part (for me) is that it contains the most recent
release it's based on (vulkan-cts-1.3.5.0 in the example above), as well
as how many commits are added on top (163) and the git hash of the last
commit (20d674342f008624b82e), which means that it contains all the
information it used to contain (not to mention the second commit hash is
still there), while also containing the useful (arguably necessary)
information of which release this comes from.

This changes the output at the start from something like this:
  dEQP Core git-20d674342f008624b82e08f26ed5572c176ba7bd (0x20d67434) starting..
to something like this:
  dEQP Core vulkan-cts-1.3.5.0-163-g20d674342f008624b82e (0x20d67434) starting..

[1] https://git-scm.com/docs/git-describe

VK-GL-CTS issue: 4327

Components: Framework

Change-Id: Ida1f315ae89ac8bf651ad7f447da05837633fa46