Fix VK/VK-SC generated file inclusion
The build system was including generated/vulkan unconditionally, which
meant that generated/vulkansc had to be pushed onto the include path
earlier everywhere it was supposed to be used. This was messy (although
it seems to have been working) and hides errors, because the
CTS_USE_VULKAN defines are propagated in the same way as the include
folders. If we only set the correct include path then these are
automatically consistent. (The generated files currently contain
CTS_USE_VULKAN{,SC} defines, but they don't build correctly if used with
the wrong ones).

One file was dependent on including both sets of generated files, and is
currently hacked to carry on doing so, but it shouldn't remain like
this.

Components: Vulkan, VulkanSC
Affects: Build only

Change-Id: I7c90b3fd733f6e9163a5e6731d6c68451492f780