Move updated fragment_density_map tests to 2023 list
Some vendors passed CTS 13_R3 but fail in CTS 13_R4/CTS 14_R1
This happens because the test contents have changed here:
https://github.com/KhronosGroup/VK-GL-CTS/commit/1de0a0106b4eab9c8ebb518a9e447dc66139db85

Now some devices are struggling to pass the new CTS. We need to allow this due to VSR constraints.

This CL moves the tests that cause issues from 2021 to 2023.

Change-Id: Ic62f6929984636ddeaea114f41e17dcfa8fdcebe
Bug: b/285981913, b/286187826