Remove allocation failed test
The allocation failed test for VK_EXT_device_memory_report tries to
make an allocation fail to test the allocation fail callback.

This is done by making a small allocation, then an allocation the size
of the heap, expecting the second to fail.

Due to platform differences, this is not a robust way force an
allocation failure, eg. the system could kill the test application
before the driver can report the failure in a low memory senario or the
driver could page out the allocation to system memory.

Since there are no guarentees in the spec on when an out of memory
error is reported and the various different handling behaviours on
different platforms, remove this test.

Affects: dEQP-VK.memory.device_memory_report.vk_device_memory.allocation_failed

Components: vulkan

VK-GL-CTS issue: 4527

Change-Id: If7d42fbfeec0f9006c2e9789391edad6d72ba8e4