Allow selecting a subset of targets at configuration time
This commit adds a single new CMake option called
"SELECTED_BUILD_TARGETS". When used and non-empty, it should contain a
list of space-separated targets to be built by default, and the effect
of setting it to a non-empty string is the same as if `--target TARGET`
had been provided one or more times on the command line.

Note the new option only applies under Windows and Unix.

No test results should be affected by this change.

Components: Framework
VK-GL-CTS issue: 3851

Change-Id: I62058c7b43a18a27d7a70c20eadc456f6c14c97f