Round ray query T value in descriptor buffer tests
The descriptor buffer tests truncate the T value of a ray query
to compare it against a target uint value.

These tests are setup in such a way that the expected T value will
always be an integer in floating-point representation.

This results in T values being truncated to (target - 1) if the
output is below the expected value due to floating-point imprecision.

Update these tests to instead round the value to the nearest integer
before truncating it. This will prevent values which are slightly below
the target from being truncated to (target - 1).

Affects: dEQP-VK.binding_model.descriptor_buffer.*

Components: Vulkan

VK-GL-CTS issue: 4587

Change-Id: Ia7475c9f3ba8bf942e5d5770f9566ae7c923a1f5