Run vk-build-programs tests in chunks
The main loop of this program was basically accumulating memory for all test, since they were initialized and stored in a collection and never deleted, until the end of execution of the program. This caused memory to go up beyond 34GB and causing several CI failures (vk-build-programs --validate-spv was killed by the system with signal -9).

This change introduces processing the tests in chunks (of 500000 at a time, but this may be changed), which allows destructions of the used programs and tests.

Memory used by this process is now less than 6GB and it runs faster (2m40s instead of 3m04s) because I removed the threads count limit too, which in my tests was not affecting memory usage.

No test results should be affected by this change.

Issue: https://gitlab.khronos.org/Tracker/vk-gl-cts/-/issues/4460

Change-Id: I6c74b98beb689d63a15090bfc3f485bdf0c82f73