Fix regression in all statistic_query tessellation tests
CL https://gerrit.khronos.org/c/vk-gl-cts/+/12373 caused a regression in all
statistic_query tessellation tests. It added a new "hasTess" member to
ParametersGraphic, which when false bypassed settings up the tessellation
shaders for the test. The problem is that hasTess was never set to true
for the tests that needed to use the tessellation shader. This resulted
in an abort from the CTS framework as it failed to lookup the shader.

Affects:

dEQP-VK.query_pool.statistics_query.*tes*

VK-GL-CTS issue: 4643

Components: Vulkan
Change-Id: Ibbc70539ee33f094d7eb99442fe34add2b4a63d1