Remove descriptions from TestNode
The descriptions alone are taking up 10s to 100s of Mbs of memory in the
background of all tests, which is increasingly problematic as the number
of tests go up. These descriptions are removed from the generic TestNode
structure and stored locally in the couple of tests that actually used
it for logging/explanation.

It's removed entirely from the XML caselists.

Components: Framework, Vulkan, OpenGL
VK-GL-CTS issue: 4681
Affects: *

Change-Id: I83c37226b5efbd253500680be31bfc84f7678015