Always wait for device idle in signal_order tests
The current code only waited for the device to be idle before tearing
down resources in case of success, but not in case of failure. This
commit adds a "deviceWaitIdle" guard so we wait in all cases before
returning from the iterate() method.

Affects:
dEQP-VK.*.signal_order.*

VK-GL-CTS issue: 4646
Components: Vulkan

Change-Id: Ic94523d766933776c850b659c674a59dbd7d99a4