Use hash instead of strings for case tree nodes to reduce mem footprint
Keeping the complete tree of test cases can be quite memory-heavy.
Especially because all the strings for all test case nodes' names are
kept in memory for the whole duration of the tests. Keeping the whole
tree in memory is necessary to be able to quickly determine if each
test needs to be run or filtered out.

This change uses a simple hash mechanism to store a size_t instead of
the strings for the node names, which is convenient if on average the
names are longer than 8(+1 for null termination) characters.

Experiments show that the size of the tree has been reduced from
420~MB to 210~MB on a 64-bit system. This can be further reduced
by using a smaller data type for the hash value (currently size_t).

While initializing the test case tree, an hash map is built to detect
collisions at runtime, and then discarded before proceeding with
the execution of the tests.

VK-GL-CTS issue: 4681

Affected tests:
dEQP-VK.*

Change-Id: I26d96dcb34b2ec828d66434fcd45a7d899e2f1a0